+++ packages/eslint-config-schibsted-chai/chai.js
@@ -1,0 +1,9 @@
+"use strict";
+
+module.exports = {
+  plugins: ["chai-friendly"],
+  rules: {
+    "no-unused-expressions": "off",
+    "chai-friendly/no-unused-expressions": "error"
+  }
+};