+++ packages/eslint-config-schibsted-node/index.js
@@ -1,0 +1,8 @@
+"use strict";
+
+module.exports = {
+  env: {
+    node: true
+  },
+  extends: ["./airbnb.js"].map(require.resolve)
+};