+++ packages/eslint-config-schibsted/CHANGELOG.md
@@ -1,8 +1,14 @@
+# 5.0.0
+- Use airbnb as initial extends
+- Use eslint-plugin-unicorn
+- Use eslint-plugin-prettier
+- Use eslint-config-prettier
+
 # 3.0.0
 - Extract all React rules into its own `react` package
 
 # 2.0.0
-- Update for ESLint 2 (@nosilleg, https://github.com/schibsted/eslint-config-spt/pull/3)
+- Update for ESLint 2 (@nosilleg, https://github.com/schibsted/eslint-config-schibsted/pull/3)
 - Change rule `strict` from option `function` to `safe` (http://eslint.org/docs/rules/strict)
 - Change rule `no-labels` from warn to error (http://eslint.org/docs/rules/no-labels)